requirements.txt
Flask
Flask-WTF
Flask-Bcrypt
Flask-Mail
flask_sqlalchemy
Jinja2
psycopg2-binary
python-dotenv
SQLAlchemy
Werkzeug
openpyxl
flask_socketio
flask_cors
alembic
Flask-Caching
redis
boto3
cryptography
pyOpenSSL
urllib3
flask_login
flask_limiter
gunicorn
Flask-Session
bleach
pydantic
celery
python-dateutil
flask_babel
beautifulsoup4
sentry-sdk
cachetools
Faker
flask_restx
pdfplumber
#spacy
itsdangerous
flask-jwt-extended
stripe
geopy==2.3.0
rapidfuzz
openai
